+++ CHANGELOG.md
@@ -9,7 +9,7 @@ please submit a pull request fixing it.
 - User profile OGP tags ([#228])
 - Added Spring 2022 RCOS proposal ([#240])
 - Updated minimum rust version to 1.58.1.
-- Fixed bug that missed verified role on users already in the RCOS Discord. ([#236], #PR)
+- Fixed bug that missed verified role on users already in the RCOS Discord. ([#236], [#242])
 
 ## 0.8.5 - December 31st, 2021
 - Changes to the config file:
@@ -180,5 +180,6 @@ please submit a pull request fixing it.
 [#220]: https://github.com/rcos/Telescope/pull/220
 [#222]: https://github.com/rcos/Telescope/pull/222
 [#228]: https://github.com/rcos/Telescope/pull/228
-[#240]: https://github.com/rcos/Telescope/pull/240
 [#236]: https://github.com/rcos/Telescope/issues/236
+[#240]: https://github.com/rcos/Telescope/pull/240
+[#242]: https://github.com/rcos/Telescope/pull/242